+++ api/models/dramaticQuestions/Maturity.cs
@@ -14,7 +14,7 @@ public class Maturity : IDramaticQuestion
     public string Negation { get { return "Immaturity perceived as maturity"; } }
     public string Positive { get { return "Maturity"; } }
 
-    public string GetLogLineContribution(int seed, string CompletionType, IGenre genre, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes)
+    public string GetLogLineContribution(long seed, string CompletionType, IGenre genre, IProblemTemplate problemTemplate, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes)
     {
         return $"The overarching theme of this story is asking if faking maturity can lead to real maturity.";
     }
